requirements.txt
facebook-business==17.0.0
python-dotenv==1.0.0
pandas==2.1.0 